% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{AndNot}
\alias{AndNot}
\alias{AndNegatedMask}
\title{AndNegatedMask query operation}
\usage{
AndNot(property, ...)

AndNegatedMask(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Same as \code{AndMask} but use the inverse of the mask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.AndNegatedMask}{documentation} for
details.
}
