% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{BeginMask}
\alias{BeginMask}
\title{BeginMask query operation}
\usage{
BeginMask(value, ...)
}
\arguments{
\item{value}{String specifying the property name for the initial mask}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Start specifying a mask to apply to an axis of the result.
Must be accompanied by an EndMask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.BeginMask}{documentation} for details.
}
