% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{BeginNegatedMask}
\alias{BeginNegatedMask}
\title{BeginNegatedMask query operation}
\usage{
BeginNegatedMask(value, ...)
}
\arguments{
\item{value}{String specifying the property name for the initial negated mask}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Start specifying a negated mask to apply to an axis of the result.
Must be accompanied by an EndMask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.BeginNegatedMask}{documentation} for details.
}
