% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete_chain}
\alias{complete_chain}
\title{Create a persistent chain from a base Daf and a new Daf}
\usage{
complete_chain(
  base_daf,
  new_daf,
  name = NULL,
  axes = NULL,
  data = NULL,
  absolute = FALSE
)
}
\arguments{
\item{base_daf}{A Daf object to use as the base (read-only) data}

\item{new_daf}{A Daf object to use as the new (writable) data on top of the base}

\item{name}{Optional name for the chained Daf object}

\item{axes}{Optional named list specifying axes to expose from the base (same format as \code{viewer})}

\item{data}{Optional named list specifying data to expose from the base (same format as \code{viewer})}

\item{absolute}{If TRUE, store the absolute path to the base_daf. If FALSE (default),
store a relative path for portability.}
}
\value{
A writable Daf object chaining the base and new data
}
\description{
Immediately after creating an empty disk-based \code{new_daf}, chain it with a disk-based
\code{base_daf} and return the new chain. If \code{axes} and/or \code{data} are specified, the \code{new_daf}
will be chained on top of a view of the \code{base_daf}.
}
\details{
This will set the \code{base_daf_repository} scalar property of the \code{new_daf} to point at the
\code{base_daf}, and if view \code{axes} or \code{data} were specified, the \code{base_daf_view} as well.
It should therefore be possible to recreate the chain by calling \code{complete_daf} in the future.

By default, the stored base path in the \code{new_daf} will be the relative path to the
\code{base_daf}, for the common case where a group of repositories is stored under a common root.
This allows the root to be renamed or moved and still allow \code{complete_daf} to work.

See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/chains.html#DataAxesFormats.Chains.complete_chain!}{documentation} for details.
}
