% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{complete_path}
\alias{complete_path}
\title{Get the complete filesystem path of a persistent Daf repository}
\usage{
complete_path(daf)
}
\arguments{
\item{daf}{A Daf object}
}
\value{
A character string with the absolute path to the persistent Daf repository,
or NULL if the repository is in-memory.
}
\description{
If the Daf repository is persistent (resides on disk), returns the absolute path leading to it.
If the repository is (at least partially) in-memory, returns NULL.
}
\details{
The returned path can be given to \code{complete_daf} to access the repository
after the current process is terminated. Note that for H5df format, the path may
end with \verb{#...} to identify a specific group inside an HDF5 file.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/readers.html#DataAxesFormats.Readers.complete_path}{documentation} for details.
}
