% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated-queries}
\alias{deprecated-queries}
\title{Deprecated query operations}
\description{
These functions have been renamed or removed in DataAxesFormats v0.2.0.
They are provided for backward compatibility but will be removed in a future version.
}
