% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{EndMask}
\alias{EndMask}
\title{EndMask query operation}
\usage{
EndMask(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Finish specifying a mask to apply to an axis, following BeginMask or BeginNegatedMask.
See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.EndMask}{documentation} for details.
}
