% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{GroupColumnsBy}
\alias{GroupColumnsBy}
\title{GroupColumnsBy query operation}
\usage{
GroupColumnsBy(value, ...)
}
\arguments{
\item{value}{String specifying the property to group columns by}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Specify value per matrix column to group the columns by. Must be followed
by a ReduceToColumn to reduce each group of columns to a single column.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.GroupColumnsBy}{documentation} for details.
}
