% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{GroupRowsBy}
\alias{GroupRowsBy}
\title{GroupRowsBy query operation}
\usage{
GroupRowsBy(value, ...)
}
\arguments{
\item{value}{String specifying the property to group rows by}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Specify value per matrix row to group the rows by. Must be followed
by a ReduceToRow to reduce each group of rows to a single row.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.GroupRowsBy}{documentation} for details.
}
