% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{has_query}
\alias{has_query}
\title{Check if a query can be applied to a Daf object}
\usage{
has_query(daf, query)
}
\arguments{
\item{daf}{A Daf object}

\item{query}{Query string or object. Can be created using query operations such as
Axis(), LookupVector(), IsGreater(), etc.}
}
\value{
TRUE if query can be applied, FALSE otherwise
}
\description{
Determines whether a query can be validly applied to a Daf object. This is useful for
checking if the properties referenced in a query exist in the Daf object before
attempting to execute the query.
}
\details{
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Operations.has_query}{documentation} for details.
}
\seealso{
Axis, LookupVector, LookupScalar, LookupMatrix, Names, AsAxis, IfMissing, IfNot,
BeginMask, BeginNegatedMask, EndMask, SquareColumnIs, SquareRowIs,
AndMask, AndNegatedMask, OrMask, OrNegatedMask, XorMask, XorNegatedMask,
IsEqual, IsNotEqual, IsLess, IsLessEqual, IsGreater, IsGreaterEqual,
IsMatch, IsNotMatch, CountBy, GroupBy, GroupColumnsBy, GroupRowsBy,
ReduceToColumn, ReduceToRow and other query operations.
}
