% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{IfMissing}
\alias{IfMissing}
\title{IfMissing query operation}
\usage{
IfMissing(default_value, ...)
}
\arguments{
\item{default_value}{Value to use when data is missing the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation providing a default value to use if the data is missing some property.
This is useful when querying for properties that might not exist for all entries, allowing you to
provide a fallback value instead of getting an error. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.IfMissing}{documentation}
for details.
}
