% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{LookupMatrix}
\alias{LookupMatrix}
\title{LookupMatrix query operation}
\usage{
LookupMatrix(value, ...)
}
\arguments{
\item{value}{String specifying the matrix property name to look up}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation for looking up the value of a matrix property.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.LookupMatrix}{documentation}
for details.
}
