% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{Lookup}
\alias{Lookup}
\alias{LookupVector}
\title{LookupVector query operation}
\usage{
Lookup(property, ...)

LookupVector(value = NULL, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}

\item{value}{String specifying the property name to look up}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation for looking up the value of a vector property with a specific name.
This operation retrieves the data associated with the specified property for the current axis context.
It is typically used after an \code{Axis} operation to select a vector property from that axis.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.LookupVector}{documentation}
for details.
}
\section{Functions}{
\itemize{
\item \code{Lookup()}: Deprecated: use \code{\link[=LookupVector]{LookupVector()}} instead.

}}
