% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{MaskSlice}
\alias{MaskSlice}
\title{MaskSlice (Deprecated)}
\usage{
MaskSlice(axis, ...)
}
\arguments{
\item{axis}{An axis to slice.}

\item{...}{Additional arguments (ignored).}
}
\description{
This function has been removed in DataAxesFormats v0.2.0.
Use \code{\link{BeginMask}}/\code{\link{EndMask}} instead.
}
