% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{Names}
\alias{Names}
\title{Names query operation}
\usage{
Names(...)
}
\arguments{
\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object that can be used in a query sequence
}
\description{
A query operation for looking up a set of names in a Daf object.
This operation retrieves the entry names of the current axis.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.Names}{documentation}
for details.
}
