% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{OrNot}
\alias{OrNot}
\alias{OrNegatedMask}
\title{OrNegatedMask query operation}
\usage{
OrNot(property, ...)

OrNegatedMask(value, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}

\item{value}{String specifying the property}
}
\value{
A query operation object
}
\description{
Same as \code{OrMask} but use the inverse of the mask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.OrNegatedMask}{documentation} for details.
}
\section{Functions}{
\itemize{
\item \code{OrNot()}: Deprecated: use \code{\link[=OrNegatedMask]{OrNegatedMask()}} instead.

}}
