% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ReduceToColumn}
\alias{ReduceToColumn}
\title{ReduceToColumn query operation}
\usage{
ReduceToColumn(reduction, ...)
}
\arguments{
\item{reduction}{A reduction operation (e.g., from Mean(), Sum(), etc.) or a string describing the reduction}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Specify a reduction operation to convert each row of a matrix to a single value,
reducing the matrix to a single column.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.ReduceToColumn}{documentation} for details.
}
