% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ReduceToRow}
\alias{ReduceToRow}
\title{ReduceToRow query operation}
\usage{
ReduceToRow(reduction, ...)
}
\arguments{
\item{reduction}{A reduction operation (e.g., from Mean(), Sum(), etc.) or a string describing the reduction}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Specify a reduction operation to convert each column of a matrix to a single value,
reducing the matrix to a single row.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.ReduceToRow}{documentation} for details.
}
