% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{SquareMaskColumn}
\alias{SquareMaskColumn}
\alias{SquareColumnIs}
\title{SquareColumnIs query operation}
\usage{
SquareMaskColumn(value, ...)

SquareColumnIs(value, ...)
}
\arguments{
\item{value}{String specifying the value}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Used when the mask matrix is square and we'd like to use a column as a mask. See the
Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.SquareColumnIs}{documentation}
for details.
}
\section{Functions}{
\itemize{
\item \code{SquareMaskColumn()}: Deprecated: use \code{\link[=SquareColumnIs]{SquareColumnIs()}} instead.

}}
