% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{SquareMaskRow}
\alias{SquareMaskRow}
\alias{SquareRowIs}
\title{SquareRowIs query operation}
\usage{
SquareMaskRow(value, ...)

SquareRowIs(value, ...)
}
\arguments{
\item{value}{String specifying the value}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Used when the mask matrix is square and we'd like to use a row as a mask. See the
Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.SquareRowIs}{documentation} for
details.
}
\section{Functions}{
\itemize{
\item \code{SquareMaskRow()}: Deprecated: use \code{\link[=SquareRowIs]{SquareRowIs()}} instead.

}}
