% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daf.R
\name{[.Daf}
\alias{[.Daf}
\title{Extract results from a Daf object using a query}
\usage{
\method{[}{Daf}(x, i, ...)
}
\arguments{
\item{x}{A Daf object}

\item{i}{A query string or object}

\item{...}{Ignored. Present for compatibility with the \code{[} generic.}
}
\value{
The result of the query
}
\description{
Extract results from a Daf object using a query
}
\details{
The expression \code{daf[query]} is equivalent to \code{get_query(daf, query, cache = TRUE)}.
See the Julia \href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Operations.get_query}{documentation} for details.
}
\examples{
\dontrun{
setup_daf()
daf <- memory_daf("example")
add_axis(daf, "cell", c("A", "B", "C"))
set_vector(daf, "cell", "score", c(1.0, 2.0, 3.0))
daf[Axis("cell") |> LookupVector("score")]
}
}
