% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{Xor}
\alias{Xor}
\alias{XorMask}
\title{XorMask query operation}
\usage{
Xor(property, ...)

XorMask(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
A query operation for flipping the set of entries of an \code{Axis}. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.XorMask}{documentation} for details.
}
