% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/query.R
\name{XorNot}
\alias{XorNot}
\alias{XorNegatedMask}
\title{XorNegatedMask query operation}
\usage{
XorNot(property, ...)

XorNegatedMask(property, ...)
}
\arguments{
\item{property}{String specifying the property}

\item{...}{Additional arguments needed to support usage of pipe operator}
}
\value{
A query operation object
}
\description{
Same as \code{XorMask} but use the inverse of the mask. See the Julia
\href{https://tanaylab.github.io/DataAxesFormats.jl/v0.2.0/queries.html#DataAxesFormats.Queries.XorNegatedMask}{documentation} for details.
}
